+++ Box.h
@@ -19,8 +19,8 @@ namespace UI
     public:
         void setLayoutMode(LayoutMode _layoutMode) { layoutMode = _layoutMode; };
         void render(SDL_Renderer *renderer, TextManager *textManager, Context *ctx) override;
-        void measure(TextManager *textManager, int availableWidth, int availableHeight) override;
-        void layoutChildrenEvenly(int i, int j, int startingX, int startingY, int availableWidth);
+        void measure(TextManager *textManager, float availableWidth, float availableHeight) override;
+        void layoutChildrenEvenly(int i, int j, float startingX, float startingY, float availableWidth);
         void setBounds(Bounds _bounds) override;
 
         void setPaddingX(int _p) { style.paddingX = _p; };
